Derived FPGA_Device from STI_Device and implemented double buffering
* Changed STI_Device::SynchronousEvent so that ::loadEvent() no longer returns the time until next available load attempt.  Now all double buffering is handled by continuously polling the FPGA.

* Wrote FPGA_Device abstract base class. This is a partial implementation of STI_Device for the case of the FPGA boards.  It overrides the default loading and playing behavior to allow for double buffering.

* Wrote FPGA_RAM_Block class to encapsulate the interaction of each FPGA_Device with the dedicated RAM chip.  This class holds the start and end RAM addresses assigned to the FPGA_Device.

* Added measurement functionality.  When devices on the FPGA acquire data, they write the result into the RAM address that previously stored the measurement event 'value'.  Each SynchronousEvent that corresponds to a measurement now contains one or more pointes to the appropriate ParsedMeasurement objects (stored in the measurements vector).  These pointers must be assigned in STI_Device::parseDeviceEvents(...) in a device-specific way.  Each event may contain multiple ParsedMeasurement objects if multiple channels are scheduled to be measured at the same time.  (To be added: throw an error during parsing if the STI_Device implementation fails to assign each ParsedMeasurement to one (and only one) SynchronousEvent.)

* Modified partner device registration to allow for devices to register themselves as partners of another device.  Previously, a device needed to define all required partners in STI_Device::definePartnerDevices().  Now the server does not forbid partners that where never defined on the device -- it only removes any dead devices from each device's list of registered partners.  This allows a device to communicate with partners that are not hard-coded in.  For example, each FPGA_Device is partners with the (still un-written) RAM controller device, and they also register themselves as a partner of the RAM controller.  This way the RAM controller can talk to any FPGA_Device that happens to be present, and it doesn't have to be told a specific list of partners in its ::definePartnerDevices().  Needs testing.

* Added double buffering to FPGA_Device::loadDeviceEvents(). This includes a preliminary automatic RAM allocation step that relies on communication of each device with a RAM controller.  Still needs testing.

* Added STI_Device::AttributeUpdater nested class that allows subclasses (such as FPGA_Device) to define and update attributes without implementing the STI_Device::updateAttribute(...) pure virtual function.  FPGA_Device implements a subclass of AttributeUpdater and overrides AttributeUpdater::updateAttribute(...).  It then adds an instance of this class using addAttributeUpdater(...).  STI_Device maintains a list of AttributeUpdater objects and calls updateAttribute when appropriate.  This keeps the interface the same for the FPGA_Device subclass.

Jason Hogan



git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti-main/branches/sti@333 8bcac300-4d60-4aec-84ba-b7489f70e69b

+++ device/src/StreamingBuffer.cpp
@@ -137,7 +137,7 @@ bool StreamingBuffer::measurementLoop()
 	// Measurement
 	ParsedMeasurement temp(tMeasurement, 0);
 	sti_Device->readChannel( temp  );
-	tMeasurement.time = getCurrentTime();
+	tMeasurement.time = static_cast<double>(getCurrentTime());
 	buffer.push_back(tMeasurement);
 
 //	cout <<  (buffer.back().time/1e9) << "\t" << (t_error/1e9) << "\t" << (t_sleep/1e9) << "\t" << busy << endl;
